+++ examples/opentelemetry-error.rs
@@ -92,17 +92,14 @@ fn main() -> Result<(), Box<dyn StdError + Send + Sync + 'static>> {
 struct WriterExporter;
 
 impl SpanExporter for WriterExporter {
-    fn export(
-        &mut self,
-        batch: Vec<sdk::trace::SpanData>,
-    ) -> futures_util::future::BoxFuture<'static, OTelSdkResult> {
+    async fn export(&self, batch: Vec<sdk::trace::SpanData>) -> OTelSdkResult {
         let mut writer = std::io::stdout();
         for span in batch {
             writeln!(writer, "{}", SpanData(span)).unwrap();
         }
         writeln!(writer).unwrap();
 
-        Box::pin(async move { OTelSdkResult::Ok(()) })
+        OTelSdkResult::Ok(())
     }
 }